Update Proxy.py
My code was encountering issues with the port connections as it kept saying the port was busy or in use. I managed to find a way around it and in the process I also tested to see if the proxy server finds a cache hit and I also got the code to create a socket to connect to the origin server.